Extract session flow.
